feat:handlerのいらないファイルを削除
