+++ api/internal/article/handler/create_article.go
@@ -5,11 +5,28 @@ import (
 	"errors"
 	"net/http"
 	"strings"
+	"time"
 	"unicode/utf8"
 
 	"github.com/Baymax6s/KOBE-Tech/api/internal/auth"
 	"github.com/gin-gonic/gin"
 )
+
+type createArticleRequest struct {
+	Title    string   `json:"title"`
+	Content  string   `json:"content"`
+	TagNames []string `json:"tags" minLength:"1" maxLength:"10"`
+} // @name server.createArticleRequest
+
+type CreateArticleJSONResponse struct {
+	ID        int64            `json:"id" binding:"required"`
+	Title     string           `json:"title" binding:"required"`
+	Content   string           `json:"content" binding:"required"`
+	UserID    int64            `json:"user_id" binding:"required"`
+	Tags      []ArticleTagJSON `json:"tags" binding:"required"`
+	CreatedAt time.Time        `json:"created_at" binding:"required"`
+	UpdatedAt time.Time        `json:"updated_at" binding:"required"`
+} // @name server.createArticleResponse
 
 // createArticleHandler godoc
 //
