+++ api/internal/article/handler/get_article.go
@@ -6,9 +6,26 @@ import (
 	"errors"
 	"net/http"
 	"strconv"
+	"time"
 
 	"github.com/gin-gonic/gin"
 )
+
+type AuthorJSON struct {
+	ID   int64  `json:"id" binding:"required"`
+	Name string `json:"name" binding:"required"`
+} // @name server.articleAuthorJSONResponse
+
+type GetArticleJSONResponse struct {
+	ID         int64            `json:"id" binding:"required"`
+	Title      string           `json:"title" binding:"required"`
+	Content    string           `json:"content" binding:"required"`
+	Author     AuthorJSON       `json:"author" binding:"required"`
+	Tags       []ArticleTagJSON `json:"tags" binding:"required"`
+	CreatedAt  time.Time        `json:"created_at" binding:"required"`
+	UpdatedAt  time.Time        `json:"updated_at" binding:"required"`
+	LikesCount int64            `json:"likes_count" binding:"required"`
+} // @name server.getArticleJSONResponse
 
 // getArticleHandler godoc
 //
