+++ api/internal/reply/handler/get_list_replies.go
@@ -6,9 +6,31 @@ import (
 	"log"
 	"net/http"
 	"strconv"
+	"time"
 
+	"github.com/Baymax6s/KOBE-Tech/api/internal/reply"
 	"github.com/gin-gonic/gin"
 )
+
+type ReplyJSON struct {
+	ID        int64     `json:"id" binding:"required"`
+	ArticleID int64     `json:"article_id" binding:"required"`
+	ParentID  *int64    `json:"parent_id,omitempty"`
+	Kind      string    `json:"kind" binding:"required" enums:"comment,question,answer"`
+	Body      string    `json:"body" binding:"required"`
+	UserID    int64     `json:"user_id" binding:"required"`
+	UserName  string    `json:"user_name" binding:"required"`
+	CreatedAt time.Time `json:"created_at" binding:"required"`
+	UpdatedAt time.Time `json:"updated_at" binding:"required"`
+} // @name server.replyJSONResponse
+
+type ListRepliesJSONResponse struct {
+	Replies []ReplyJSON `json:"replies" binding:"required"`
+} // @name server.listRepliesResponse
+
+type ErrorResponse struct {
+	Message string `json:"message"`
+} // @name server.replyErrorResponse
 
 // listRepliesHandler godoc
 //
@@ -49,4 +71,26 @@ func (h *Handler) ListReplies(ctx context.Context, articleID int64) (ListReplies
 	}
 
 	return ListRepliesJSONResponse{Replies: newReplyJSONs(replies)}, nil
+}
+
+func newReplyJSONs(replies []reply.Reply) []ReplyJSON {
+	items := make([]ReplyJSON, 0, len(replies))
+	for _, item := range replies {
+		items = append(items, newReplyJSON(item))
+	}
+	return items
+}
+
+func newReplyJSON(item reply.Reply) ReplyJSON {
+	return ReplyJSON{
+		ID:        item.ID,
+		ArticleID: item.ArticleID,
+		ParentID:  item.ParentID,
+		Kind:      item.Kind.String(),
+		Body:      item.Body,
+		UserID:    item.UserID,
+		UserName:  item.UserName,
+		CreatedAt: item.CreatedAt,
+		UpdatedAt: item.UpdatedAt,
+	}
 }
